Preserve terminal recovery with window docks
